Major API changes
Moved the file into the PDF object, to allow for future incremental
output.
Started a new graphics context struct, to create a better object.
Setting color, and other graphics parameters should be part of the
graphics context, OR the individual graphics objects. Perhaps a
macro to implement a graphics context?